feat: list all items paginated

+++ domain-model/models/src/commonMain/kotlin/dev/programadorthi/models/PaginatedItem.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 public data class PaginatedItem<T>(
-    val index: Int,
+    val index: Long,
     val size: Int,
-    val count: Int,
+    val count: Long,
     val data: List<T>,
 )
